# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Forex Trading Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pairs/{symbol}:
    get:
      operationId: getPairDetail
      tags: [pairs]
      summary: Get detailed data for a specific pair including candles and S&R zones
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pair detail with candles and zones
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairDetail"
        "404":
          description: Pair not found
...
components:
  schemas:
...
    Candle:
      type: object
      required:
        - timestamp
        - open
        - high
        - low
        - close
        - volume
      properties:
        timestamp:
          type: integer
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
    FairValueGap:
      type: object
      required:
        - top
        - bottom
        - isBullish
        - timestamp
        - filled
      properties:
        top:
          type: number
        bottom:
          type: number
        isBullish:
          type: boolean
        timestamp:
          type: integer
        filled:
          type: boolean
    MarketStructureInfo:
      type: object
      required:
        - trend
      properties:
        trend:
          type: string
          enum: [up, down, range]
        lastSwingHigh:
          type: ["number", "null"]
        lastSwingLow:
          type: ["number", "null"]
        bos:
          type: ["object", "null"]
          properties:
            price:
              type: number
            direction:
              type: string
              enum: [up, down]
            timestamp:
              type: integer
        choch:
          type: ["object", "null"]
          properties:
            price:
              type: number
            direction:
              type: string
              enum: [up, down]
            timestamp:
              type: integer
    VpBin:
      type: object
      required:
        - priceMid
        - normalizedVol
        - isPoc
        - isValueArea
      properties:
        priceMid:
          type: number
        normalizedVol:
          type: number
        isPoc:
          type: boolean
        isValueArea:
          type: boolean
    VolumeProfile:
      type: object
      required:
        - poc
        - vah
        - val
        - bins
      properties:
        poc:
          type: number
          description: "Point of Control — price level with highest volume"
        vah:
          type: number
          description: "Value Area High (70% of volume)"
        val:
          type: number
          description: "Value Area Low (70% of volume)"
        bins:
          type: array
          items:
            $ref: "#/components/schemas/VpBin"
    OrderBlock:
      type: object
      required:
        - top
        - bot
        - isBullish
        - timestamp
        - mitigated
      properties:
        top:
          type: number
        bot:
          type: number
        isBullish:
          type: boolean
          description: "Bullish OB = potential buy zone; Bearish OB = potential sell zone"
        timestamp:
          type: integer
        mitigated:
          type: boolean
          description: "True if price has already returned through the OB"
    LiquidityGrab:
      type: object
      required:
        - price
        - isBuySide
        - timestamp
        - wickExtent
      properties:
        price:
          type: number
          description: "The swing level that was swept"
        isBuySide:
          type: boolean
          description: "True = buy-side liquidity was grabbed (wick above swing high)"
        timestamp:
          type: integer
        wickExtent:
          type: number
    Zone:
      type: object
      required:
        - top
        - bot
        - center
        - isResistance
        - touches
        - strength
        - rating
        - htfConfluence
      properties:
        top:
          type: number
        bot:
          type: number
        center:
          type: number
        isResistance:
          type: boolean
        touches:
          type: integer
        strength:
          type: number
        volSum:
          type: ["number", "null"]
        firstBar:
          type: integer
        lastTouch:
          type: integer
        rating:
          type: integer
        htfConfluence:
          type: boolean
        htfLevel:
          type: string
          enum: [none, H1, H4, H1_H4]
        probabilityScore:
          type: ["integer", "null"]
        nearRoundNumber:
          type: ["boolean", "null"]
        ageBars:
          type: ["integer", "null"]
    PairDetail:
      type: object
      required:
        - symbol
        - displayName
        - currentPrice
        - bid
        - ask
        - change
        - changePct
        - zoneStatus
        - signal
        - candles
        - resistanceZones
        - supportZones
        - updatedAt
      properties:
        symbol:
          type: string
        displayName:
          type: string
        currentPrice:
          type: number
        bid:
          type: number
        ask:
          type: number
        change:
          type: number
        changePct:
          type: number
        zoneStatus:
          type: string
          enum: [neutral, resistance, support, near_resistance, near_support]
        signal:
          type: string
          enum: [none, short, long]
        pattern:
          type: string
          enum: [none, pin_bar_bullish, pin_bar_bearish, engulfing_bullish, engulfing_bearish, doji_bearish, doji_bullish, doji]
        nearestResistance:
          type: ["number", "null"]
        nearestSupport:
          type: ["number", "null"]
        candles:
          type: array
          items:
            $ref: "#/components/schemas/Candle"
        resistanceZones:
          type: array
          items:
            $ref: "#/components/schemas/Zone"
        supportZones:
          type: array
          items:
            $ref: "#/components/schemas/Zone"
        adrPips:
          type: ["number", "null"]
        todayRangePips:
          type: ["number", "null"]
        adrPercent:
          type: ["number", "null"]
        adrRisk:
          type: string
          enum: [low, medium, high, very_high, unknown]
        dailyBias:
          type: string
          enum: [bullish, bearish, neutral]
        ema50:
          type: ["number", "null"]
        ema200:
          type: ["number", "null"]
        ema50Values:
          type: array
          items:
            type: number
        ema200Values:
          type: array
          items:
            type: number
        trend:
          type: string
          enum: [up, down, range]
        fairValueGaps:
          type: array
          items:
            $ref: "#/components/schemas/FairValueGap"
        marketStructure:
          $ref: "#/components/schemas/MarketStructureInfo"
        psychologicalLevels:
          type: array
          items:
            type: number
        volumeProfile:
          $ref: "#/components/schemas/VolumeProfile"
        orderBlocks:
          type: array
          items:
            $ref: "#/components/schemas/OrderBlock"
        liquidityGrabs:
          type: array
          items:
            $ref: "#/components/schemas/LiquidityGrab"
        updatedAt:
          type: string
...