# lib/api-zod/src/generated/types/liquidityGrab.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Forex Trading Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface LiquidityGrab {
  /** The swing level that was swept */
  price: number;
  /** True = buy-side liquidity was grabbed (wick above swing high) */
  isBuySide: boolean;
  timestamp: number;
  wickExtent: number;
}