# lib/api-zod/src/generated/types/orderBlock.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Forex Trading Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderBlock {
  top: number;
  bot: number;
  /** Bullish OB = potential buy zone; Bearish OB = potential sell zone */
  isBullish: boolean;
  timestamp: number;
  /** True if price has already returned through the OB */
  mitigated: boolean;
}