# lib/api-zod/src/generated/types/pairDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Forex Trading Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Candle } from './candle';
import type { FairValueGap } from './fairValueGap';
import type { LiquidityGrab } from './liquidityGrab';
import type { MarketStructureInfo } from './marketStructureInfo';
import type { OrderBlock } from './orderBlock';
import type { PairDetailAdrRisk } from './pairDetailAdrRisk';
import type { PairDetailDailyBias } from './pairDetailDailyBias';
import type { PairDetailPattern } from './pairDetailPattern';
import type { PairDetailSignal } from './pairDetailSignal';
import type { PairDetailTrend } from './pairDetailTrend';
import type { PairDetailZoneStatus } from './pairDetailZoneStatus';
import type { VolumeProfile } from './volumeProfile';
import type { Zone } from './zone';

export interface PairDetail {
  symbol: string;
  displayName: string;
  currentPrice: number;
  bid: number;
  ask: number;
  change: number;
  changePct: number;
  zoneStatus: PairDetailZoneStatus;
  signal: PairDetailSignal;
  pattern?: PairDetailPattern;
  /** @nullable */
  nearestResistance?: number | null;
  /** @nullable */
  nearestSupport?: number | null;
  candles: Candle[];
  resistanceZones: Zone[];
  supportZones: Zone[];
  /** @nullable */
  adrPips?: number | null;
  /** @nullable */
  todayRangePips?: number | null;
  /** @nullable */
  adrPercent?: number | null;
  adrRisk?: PairDetailAdrRisk;
  dailyBias?: PairDetailDailyBias;
  /** @nullable */
  ema50?: number | null;
  /** @nullable */
  ema200?: number | null;
  ema50Values?: number[];
  ema200Values?: number[];
  trend?: PairDetailTrend;
  fairValueGaps?: FairValueGap[];
  marketStructure?: MarketStructureInfo;
  psychologicalLevels?: number[];
  volumeProfile?: VolumeProfile;
  orderBlocks?: OrderBlock[];
  liquidityGrabs?: LiquidityGrab[];
  updatedAt: string;
}