# lib/api-zod/src/generated/types/volumeProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Forex Trading Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { VpBin } from './vpBin';

export interface VolumeProfile {
  /** Point of Control — price level with highest volume */
  poc: number;
  /** Value Area High (70% of volume) */
  vah: number;
  /** Value Area Low (70% of volume) */
  val: number;
  bins: VpBin[];
}